backend/requirements.txt
Flask
Flask-SQLAlchemy  # If using SQLAlchemy
Flask-Mail
Flask-CORS
Flask-Login
psycopg2-binary   # If using PostgreSQL
python-dotenv
web3
requests
pyotp           # For OTP generation
google-auth
google-auth-oauthlib
itsdangerous
gunicorn
eth-account
erc6492-signature-verifier
py-solc-x
eth-tester[py-evm]